// apps/api/src/modules/organizations/user-role.enum.ts
export enum UserRole {
  ADMIN = 'admin',
  COMMERCIAL_DIRECTOR = 'commercial_director',
  CREATIVE_DIRECTOR = 'creative_director',
  OPERATIONS_DIRECTOR = 'operations_director',
  ART_DIRECTOR = 'art_director',
  AV_DIRECTOR = 'av_director',
  AI_LEAD = 'ai_lead',
  COMMUNITY_MANAGER = 'community_manager',
  DESIGNER = 'designer',
  AUDIOVISUAL = 'audiovisual',
  CLIENT = 'client',
}